Separate part of vgmstream.c to decode.c and render.c
