+++ src/layout/blocked.c
@@ -1,5 +1,6 @@
 #include "layout.h"
 #include "../vgmstream.h"
+#include "../decode.h"
 
 
 /* Decodes samples for blocked streams.
