+++ src/render.h
@@ -1,0 +1,10 @@
+#ifndef _RENDER_H
+#define _RENDER_H
+
+#include "vgmstream.h"
+
+void free_layout(VGMSTREAM* vgmstream);
+void reset_layout(VGMSTREAM* vgmstream);
+
+
+#endif
