+++ src/vgmstream.h
@@ -1137,28 +1137,10 @@ int vgmstream_is_virtual_filename(const char* filename);
 /* -------------------------------------------------------------------------*/
 
 /* Allocate initial memory for the VGMSTREAM */
-VGMSTREAM * allocate_vgmstream(int channel_count, int looped);
+VGMSTREAM* allocate_vgmstream(int channel_count, int looped);
 
 /* Prepare the VGMSTREAM's initial state once parsed and ready, but before playing. */
 void setup_vgmstream(VGMSTREAM* vgmstream);
-
-/* Get the number of samples of a single frame (smallest self-contained sample group, 1/N channels) */
-int get_vgmstream_samples_per_frame(VGMSTREAM* vgmstream);
-/* Get the number of bytes of a single frame (smallest self-contained byte group, 1/N channels) */
-int get_vgmstream_frame_size(VGMSTREAM* vgmstream);
-/* In NDS IMA the frame size is the block size, so the last one is short */
-int get_vgmstream_samples_per_shortframe(VGMSTREAM* vgmstream);
-int get_vgmstream_shortframe_size(VGMSTREAM* vgmstream);
-
-/* Decode samples into the buffer. Assume that we have written samples_written into the
- * buffer already, and we have samples_to_do consecutive samples ahead of us. */
-void decode_vgmstream(VGMSTREAM* vgmstream, int samples_written, int samples_to_do, sample_t* buffer);
-
-/* Calculate number of consecutive samples to do (taking into account stopping for loop start and end) */
-int get_vgmstream_samples_to_do(int samples_this_block, int samples_per_frame, VGMSTREAM* vgmstream);
-
-/* Detect loop start and save values, or detect loop end and restore (loop back). Returns 1 if loop was done. */
-int vgmstream_do_loop(VGMSTREAM* vgmstream);
 
 /* Open the stream for reading at offset (taking into account layouts, channels and so on).
  * Returns 0 on failure */
@@ -1170,7 +1152,5 @@ void get_vgmstream_coding_description(VGMSTREAM* vgmstream, char* out, size_t ou
 void get_vgmstream_layout_description(VGMSTREAM* vgmstream, char* out, size_t out_size);
 void get_vgmstream_meta_description(VGMSTREAM* vgmstream, char* out, size_t out_size);
 
-
-int render_fade(VGMSTREAM* vgmstream, sample_t* buf, int samples_done);
 void setup_state_vgmstream(VGMSTREAM* vgmstream);
 #endif
